refactor: move ops util function out of tx.ts
