+++ packages/app/src/systems/Transaction/utils/index.tsx
@@ -6,3 +6,4 @@ export * from './tx.types';
 export * from './icon';
 export * from './text';
 export * from './date';
+export * from './operation';
